requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.2
chex==0.1.90
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dotmap==1.3.30
etils==1.13.0
evosax @ git+https://github.com/RobertTLange/evosax.git@e357db0fd93fa613d0bdfed12160a200bb587d2c
filelock==3.13.1
flax==0.11.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.3
numpy==2.3.2
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.20
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
SciencePlots==2.1.1
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.34.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tensorstore==0.1.76
threadpoolctl==3.6.0
toolz==1.0.0
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
treescope==0.1.9
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wandb==0.21.0
zipp==3.23.0
